add GradieteModel, Rename View Controllers
